+++ L2J_Mobius_C6_Interlude/java/com/l2jmobius/gameserver/instancemanager/FourSepulchersManager.java
@@ -1605,7 +1605,7 @@ public class FourSepulchersManager extends GrandBossManager
 	
 	protected byte minuteSelect(byte min)
 	{
-		if (((double) min % 5) != 0)// if doesn't divides on 5 fully
+		if ((min % 5) != 0)// if doesn't divides on 5 fully
 		{
 			// mad table for selecting proper minutes...
 			// may be there is a better way to do this
